Example working with animation loop
