+++ examples/showcase/geospatial/package.json
@@ -5,8 +5,9 @@
     "build": "webpack --env.local --env.analyze --profile --json > stats.json"
   },
   "dependencies": {
-    "@luma.gl/core": "^7.2.0",
-    "@luma.gl/webgl": "^7.2.0",
+    "@luma.gl/gltools": "^8.0.0-beta.1",
+    "@luma.gl/engine": "^8.0.0-beta.1",
+    "@luma.gl/webgl": "^8.0.0-beta.1",
     "mapbox-gl": "^1.2.1"
   },
   "devDependencies": {
