+++ examples/showcase/geospatial/webpack.config.js
